Reorganize module structure
- Move Account and Portfolio to /trading
- Separate commission models from calculators
