+++ nautilus_trader/analysis/performance.pyx
@@ -34,11 +34,11 @@ import pandas as pd
 from scipy.stats import kurtosis
 from scipy.stats import skew
 
-from nautilus_trader.common.account cimport Account
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.model.events cimport AccountState
 from nautilus_trader.model.objects cimport Money
 from nautilus_trader.model.position cimport Position
+from nautilus_trader.trading.account cimport Account
 
 
 cdef class PerformanceAnalyzer:
