+++ nautilus_trader/execution/cache.pxd
@@ -13,7 +13,6 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
-from nautilus_trader.common.account cimport Account
 from nautilus_trader.common.logging cimport LoggerAdapter
 from nautilus_trader.execution.base cimport ExecutionCacheReadOnly
 from nautilus_trader.execution.database cimport ExecutionDatabase
@@ -25,6 +24,7 @@ from nautilus_trader.model.identifiers cimport Symbol
 from nautilus_trader.model.objects cimport Decimal
 from nautilus_trader.model.order cimport Order
 from nautilus_trader.model.position cimport Position
+from nautilus_trader.trading.account cimport Account
 from nautilus_trader.trading.strategy cimport TradingStrategy
