+++ nautilus_trader/execution/live.pyx
@@ -18,7 +18,6 @@ import threading
 
 from nautilus_trader.common.clock cimport Clock
 from nautilus_trader.common.logging cimport Logger
-from nautilus_trader.common.portfolio cimport Portfolio
 from nautilus_trader.common.uuid cimport UUIDFactory
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.message cimport Message
@@ -29,6 +28,7 @@ from nautilus_trader.model.commands cimport Command
 from nautilus_trader.model.events cimport Event
 from nautilus_trader.model.identifiers cimport AccountId
 from nautilus_trader.model.identifiers cimport TraderId
+from nautilus_trader.trading.portfolio cimport Portfolio
 
 
 cdef class LiveExecutionEngine(ExecutionEngine):
