+++ nautilus_trader/trading/portfolio.pyx
