+++ nautilus_trader/trading/trader.pxd
@@ -17,7 +17,6 @@ from nautilus_trader.analysis.performance cimport PerformanceAnalyzer
 from nautilus_trader.analysis.reports cimport ReportProvider
 from nautilus_trader.common.clock cimport Clock
 from nautilus_trader.common.logging cimport LoggerAdapter
-from nautilus_trader.common.portfolio cimport Portfolio
 from nautilus_trader.common.uuid cimport UUIDFactory
 from nautilus_trader.core.fsm cimport FiniteStateMachine
 from nautilus_trader.data.engine cimport DataEngine
@@ -25,6 +24,7 @@ from nautilus_trader.execution.engine cimport ExecutionEngine
 from nautilus_trader.model.c_enums.component_state cimport ComponentState
 from nautilus_trader.model.identifiers cimport AccountId
 from nautilus_trader.model.identifiers cimport TraderId
+from nautilus_trader.trading.portfolio cimport Portfolio
 
 
 cdef class Trader:
