+++ tests/unit_tests/common/test_common_account.py
@@ -15,12 +15,12 @@
 
 import unittest
 
-from nautilus_trader.common.account import Account
 from nautilus_trader.core.uuid import uuid4
 from nautilus_trader.model.currency import Currency
 from nautilus_trader.model.events import AccountState
 from nautilus_trader.model.identifiers import AccountId
 from nautilus_trader.model.objects import Money
+from nautilus_trader.trading.account import Account
 from tests.test_kit.stubs import UNIX_EPOCH
