+++ tests/unit_tests/trading/test_trading_trader.py
@@ -24,8 +24,6 @@ from nautilus_trader.backtest.logging import TestLogger
 from nautilus_trader.backtest.market import SimulatedMarket
 from nautilus_trader.backtest.models import FillModel
 from nautilus_trader.common.clock import TestClock
-from nautilus_trader.common.market import GenericCommissionModel
-from nautilus_trader.common.portfolio import Portfolio
 from nautilus_trader.common.uuid import TestUUIDFactory
 from nautilus_trader.execution.database import BypassExecutionDatabase
 from nautilus_trader.execution.engine import ExecutionEngine
@@ -37,6 +35,8 @@ from nautilus_trader.model.identifiers import IdTag
 from nautilus_trader.model.identifiers import StrategyId
 from nautilus_trader.model.identifiers import TraderId
 from nautilus_trader.model.identifiers import Venue
+from nautilus_trader.trading.commission import GenericCommissionModel
+from nautilus_trader.trading.portfolio import Portfolio
 from nautilus_trader.trading.trader import Trader
 from tests.test_kit.data import TestDataProvider
 from tests.test_kit.strategies import EmptyStrategy
